数据统计，管理端，完结撒花

+++ Bibibi-common/src/main/java/com/Bibibi/service/impl/VideoDanmuServiceImpl.java
@@ -148,6 +148,9 @@ public class VideoDanmuServiceImpl implements VideoDanmuService {
         if (videoInfo == null) {
             throw new BusinessException(ResponseCodeEnum.CODE_600);
         }
+        if (userId != null && !videoInfo.getUserId().equals(userId)) {
+            throw new BusinessException(ResponseCodeEnum.CODE_600);
+        }
         videoDanmuMappers.deleteByDanmuId(danmuId);
     }
 }
